requirements.txt
cachetools==5.3.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-pubsub==2.16.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.54.0
proto-plus==1.22.2
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==9.1.0
pyusb==1.2.1
rsa==4.9
